OGC Refactoring Part 4 - Move ogc models to package /ogc/model
