+++ server/src/main/java/de/ingrid/igeserver/ogc/OgcHtmlConverterService.kt
@@ -23,9 +23,8 @@ import com.fasterxml.jackson.databind.JsonNode
 import com.fasterxml.jackson.databind.node.JsonNodeType
 import com.fasterxml.jackson.databind.node.ObjectNode
 import de.ingrid.igeserver.configuration.GeneralProperties
-import de.ingrid.igeserver.model.Link
+import de.ingrid.igeserver.ogc.model.Link
 import de.ingrid.igeserver.services.QueryMetadata
-import org.springframework.beans.factory.annotation.Autowired
 import org.springframework.context.annotation.Profile
 import org.springframework.stereotype.Service
