+++ server/src/main/java/de/ingrid/igeserver/ogc/api/OgcRecordApiController.kt
@@ -22,7 +22,7 @@ package de.ingrid.igeserver.ogc.api
 import com.fasterxml.jackson.databind.JsonNode
 import de.ingrid.igeserver.api.ImportOptions
 import de.ingrid.igeserver.exports.ExporterFactory
-import de.ingrid.igeserver.model.Link
+import de.ingrid.igeserver.ogc.model.Link
 import de.ingrid.igeserver.model.ResearchResponse
 import de.ingrid.igeserver.ogc.OgcFilterParameter
 import de.ingrid.igeserver.ogc.OgcApiResearchQueryFactory
